requirements.txt
numpy==1.26.4
pandas==2.2.2
requests==2.31.0
python-dotenv==1.0.0
scikit-learn==1.5.0
xgboost==2.1.0
flask==3.0.3
matplotlib==3.9.0
seaborn==0.13.2
prophet==1.1.5
statsmodels==0.14.2
